ci(oceans-lab): add GitHub Actions workflow + Playwright e2e suite (41 tests)
New .github/workflows/oceans-ci.yml reusable workflow: build job runs
turbo build/lint/typecheck/test; e2e job runs Playwright in pinned
mcr.microsoft.com/playwright:v1.57.0-noble container with sparse
checkout and artifact upload on failure.  frontend-ci.yml wires it in
with a dorny/paths-filter oceans gate and push-trigger.

POM refactored to use a11y selectors throughout: OceansPage base with
shared locators → TrainingPage abstract with classifyOne/train/fullFlow
→ FishVTrashPage and FishShortPage concrete subclasses; yes-button
locators use exact:true to avoid strict-mode collision with "Not Fish" /
"Not <word>".  monolithic oceans.spec.ts split into five focused files:
fishvtrash, fishshort, accessibility, demo-shell, console-health.

Component a11y fixes required to make selectors work:
- train/index.tsx: erase icon-div → <button type=button aria-label=Erase>
- predict/index.tsx: media-control spans get role=button + tabIndex + aria-label
- pond/index.tsx: pond surface gets role=button + aria-label; info div → <button aria-pressed>
